+++ DoorOpenerBruh/Assets/Pieces/Doors/OtherDoors.cs
@@ -1,0 +1,24 @@
+﻿using DoorOpenerBruh.Configuration;
+
+namespace DoorOpenerBruh.Assets.Pieces.Doors;
+
+public class OtherDoors : DoorPiece
+{
+    public OtherDoors(string prefabName, string pieceName, string configSection) : base(prefabName, pieceName, configSection)
+    {
+        RegisterConfigSettings();
+    }
+
+    public override bool DoorAutomationEnabled(Door trackedDoor)
+    {
+        var enabled = ComputeAutomation(trackedDoor, true);
+
+        return enabled;
+    }
+
+    internal sealed override void RegisterConfigSettings()
+    {
+        RegisterAutomationMechanic(Components.AutomationMechanic.OpenAllDoors);
+        RegisterCheckForKey(false);
+    }
+}