+++ DoorOpenerBruh/Assets/Pieces/Doors/SunkenCryptIronGate.cs
@@ -1,0 +1,25 @@
+﻿using DoorOpenerBruh.Configuration;
+
+namespace DoorOpenerBruh.Assets.Pieces.Doors;
+
+public class SunkenCryptIronGate : DoorPiece
+{
+    public SunkenCryptIronGate(string prefabName, string pieceName, string configSection) : base(prefabName, pieceName, configSection)
+    {
+        RegisterConfigSettings();
+    }
+
+    public override bool DoorAutomationEnabled(Door trackedDoor)
+    {
+        var enabled = ComputeAutomation(trackedDoor, true) && 
+                      ConfigRegistry.OpenSunkenCryptDoors.Value;
+
+        return enabled;
+    }
+
+    internal sealed override void RegisterConfigSettings()
+    {
+        RegisterAutomationMechanic();
+        RegisterCheckForKey(false);
+    }
+}