+++ DoorOpenerBruh/Configuration/ConfigRegistry.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using BepInEx.Configuration;
-using UnityEngine;
 using Vapok.Common.Abstractions;
 using Vapok.Common.Managers.Configuration;
 using Vapok.Common.Shared;
@@ -11,6 +10,7 @@ namespace DoorOpenerBruh.Configuration
     {
         //Configuration Entry Privates
         internal static ConfigEntry<bool> OpenCryptDoors { get; private set;}
+        internal static ConfigEntry<bool> OpenSunkenCryptDoors { get; private set;}
         internal static ConfigEntry<bool> Enabled { get; private set;}
         
         public static Waiting Waiter;
@@ -34,10 +34,15 @@ namespace DoorOpenerBruh.Configuration
                     null, 
                     new ConfigurationManagerAttributes { Category = "Synced Settings", Order = 1 }));
 
-            OpenCryptDoors = SyncedConfig("Synced Settings", "Open Crypt Doors", false,
+            OpenCryptDoors = SyncedConfig("Synced Settings", "Open Burial Chamber Doors", false,
                 new ConfigDescription("If true, will auto open doors when in crypts.",
                     null, 
                     new ConfigurationManagerAttributes { Category = "Synced Settings", Order = 2 }));
+
+            OpenSunkenCryptDoors = SyncedConfig("Synced Settings", "Open Sunken Crypt Gates", false,
+                new ConfigDescription("If true, will auto open doors when in crypts.",
+                    null, 
+                    new ConfigurationManagerAttributes { Category = "Synced Settings", Order = 3 }));
         }
     }