+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.1.0 - Door Management and Config
+* Introducing Config Settings for Everydoor in the Game.
+  * Auto Close/Open settings based on type of Door
+  * Auto Close/Open Methods based on Self-made Door, or Player Only made Doors.
+  * Configuration is ServerSync-able
+* Improved the detection and increased the check speed to attempt to prevent non-closures.
+
 # v1.0.1 - Updating to work without other mods now.
 * Fixing startup errors when the mod is ran by itself.